Fix bug where calling play function with a non-numeric throws an exception.
This change allows the audioclip play function to be called with non-numeric arguments,
and have them resolve to defaults. This allows the play functions to be much more easily
used as event handlers, for example.

Prior to this change, doing something like:

```
var snare = audioclip('snare.mp3');
button.on('click', snare);
```

will result in an exception, since an event object is passed to the play function,
and the play function attempts to use that as a number (specifically, to determine
the timeout before playing the clip).